final->override (get better warnings -> fix some missing overrides)

+++ test/fuzz/joedbc_fuzzer.cpp
@@ -11,7 +11,7 @@ extern "C" int LLVMFuzzerTestOneInput(const uint8_t *Data, size_t Size)
   joedb::Readonly_Memory_File file(Data, Size);
   joedb::Readonly_Journal journal(joedb::Journal_Construction_Lock(file, true));
   my_namespace::is_nested::test::Database db;
-  db.set_max_record_id(size_t(journal.get_checkpoint()));
+  db.set_max_record_id(journal.get_checkpoint());
   journal.replay_log(db);
  }
  catch (const joedb::Exception &)
